stop deleting overlays & instead apply them properly
previously, the scripts attempted to move overlays to
a subdirectory of /system so that they could be installed
on the system partition. this is incorrect as android will
not read overlays from /system, they have to be in product
or vendor. the overlays that opengapps include are vital to
full & bug free functionality of gapps - for example, the
play store overlay included in opengapps pico is required
for some apps & parts of gms to work properly, as well as
for sign in & management of family linked or child google
accounts.